move SDDigital watchface to sub dir
