+++ apps/wear/watchfaces/SmartDriveDigital/.gradle/buildOutputCleanup/cache.properties
@@ -1,0 +1,2 @@
+#Tue Nov 19 17:45:28 CST 2019
+gradle.version=5.6.1
