+++ apps/wear/watchfaces/SmartDriveDigital/app/.gitignore
