+++ apps/wear/watchfaces/SmartDriveDigital/app/build.gradle
@@ -4,7 +4,7 @@ android {
     compileSdkVersion 28
     buildToolsVersion "28.0.3"
     defaultConfig {
-        applicationId "com.example.smartdrivemx2"
+        applicationId "com.permobil.smartdrivemx2digital"
         minSdkVersion 24
         targetSdkVersion 28
         versionCode 1
