+++ apps/wear/watchfaces/SmartDriveDigital/app/proguard-rules.pro
