+++ apps/wear/watchfaces/SmartDriveDigital/app/src/main/assets/fonts/opensans_regular.ttf
