+++ apps/wear/watchfaces/SmartDriveDigital/app/src/main/java/com/permobil/smartdrivemx2digital/ComplicationWatchFaceService.java
@@ -1,4 +1,4 @@
-package com.example.smartdrivemx2;
+package com.permobil.smartdrivemx2digital;
 
 import android.app.PendingIntent;
 import android.content.BroadcastReceiver;
