+++ apps/wear/watchfaces/SmartDriveDigital/app/src/main/res/drawable-nodpi/add_complication.png
