+++ apps/wear/watchfaces/SmartDriveDigital/app/src/main/res/drawable-nodpi/added_complication.png
