+++ apps/wear/watchfaces/SmartDriveDigital/app/src/main/res/drawable-nodpi/preview_watch_face.png
