+++ apps/wear/watchfaces/SmartDriveDigital/app/src/main/res/layout/activity_config.xml
