+++ apps/wear/watchfaces/SmartDriveDigital/app/src/main/res/mipmap-hdpi/ic_launcher.png
