+++ apps/wear/watchfaces/SmartDriveDigital/app/src/main/res/mipmap-mdpi/ic_launcher.png
