+++ apps/wear/watchfaces/SmartDriveDigital/app/src/main/res/mipmap-xxhdpi/ic_launcher.png
