+++ apps/wear/watchfaces/SmartDriveDigital/app/src/main/res/values/color.xml
