+++ apps/wear/watchfaces/SmartDriveDigital/app/src/main/res/values/dimens.xml
