+++ apps/wear/watchfaces/SmartDriveDigital/app/src/main/resources/sentry.properties
