+++ apps/wear/watchfaces/SmartDriveDigital/build.gradle
