+++ apps/wear/watchfaces/SmartDriveDigital/gradle.properties
