+++ apps/wear/watchfaces/SmartDriveDigital/settings.gradle
